Created separate file for classes (person,display)
I took some of the concepts from the parcel-kobe branch and added it to
this one. I also added button elements to our display and
cleaned up the main.html/index.html (whatever we're calling it) so that
the event listeners are applied to the display object's buttons. This
allows the use of anonymous functions inside the event listener which
gives more space in main.html. Button elements are no longer needed in
html.